requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
werkzeug
pyodbc
